Move old scripts into dedicated scripts dir
