+++ scripts/getTaunts.js
